+++ ECHO/Core/Classes/SocketOperation/GetAccountDepositsSocketOperation.swift
@@ -11,51 +11,50 @@ import Foundation
 /**
     Returns all approved deposits, for the given account id
  
-    - Return: [[DepositEth]](DepositEth)
+    - Return: [[EthDeposit]](EthDeposit)
  */
 struct GetAccountDepositsSocketOperation: SocketOperation {
-    
     var method: SocketOperationType
     var operationId: Int
     var apiId: Int
     var accountId: String
-    var completion: Completion<[DepositEth]>
+    var type: SidechainType?
+    var completion: Completion<[SidechainDepositEnum]>
     
     func createParameters() -> [Any] {
         let array: [Any] = [apiId,
                             SocketOperationKeys.getAccountDeposits.rawValue,
-                            [accountId]]
+                            [accountId, type?.rawValue ?? String()]]
         return array
     }
     
     func handleResponse(_ response: ECHODirectResponse) {
         
         do {
-            
             switch response.response {
             case .error(let error):
-                let result = Result<[DepositEth], ECHOError>(error: ECHOError.internalError(error.message))
+                let result = Result<[SidechainDepositEnum], ECHOError>(error: ECHOError.internalError(error.message))
                 completion(result)
             case .result(let result):
-                
                 switch result {
                 case .array(let array):
                     let data = try JSONSerialization.data(withJSONObject: array, options: [])
-                    let accountIds = try JSONDecoder().decode([DepositEth].self, from: data)
-                    let result = Result<[DepositEth], ECHOError>(value: accountIds)
+                    let deposits = try JSONDecoder().decode([SidechainDepositEnum].self, from: data)
+                    let result = Result<[SidechainDepositEnum], ECHOError>(value: deposits)
                     completion(result)
+                    
                 default:
                     throw ECHOError.encodableMapping
                 }
             }
         } catch {
-            let result = Result<[DepositEth], ECHOError>(error: ECHOError.encodableMapping)
+            let result = Result<[SidechainDepositEnum], ECHOError>(error: ECHOError.encodableMapping)
             completion(result)
         }
     }
     
     func forceEnd(error: ECHOError) {
-        let result = Result<[DepositEth], ECHOError>(error: error)
+        let result = Result<[SidechainDepositEnum], ECHOError>(error: ECHOError.connectionLost)
         completion(result)
     }
 }