+++ ECHO/Core/Facades/Implementations/ContractsFacadeImp.swift
@@ -60,7 +60,7 @@ final public class ContractsFacadeImp: ContractsFacade, ECHOQueueble {
         noticeDelegateHandler.delegate = self
     }
     
-    public func getContractLogs(contractId: String, fromBlock: Int, limit: Int, completion: @escaping Completion<[ContractLogEnum]>) {
+    public func getContractLogs(contractId: String, fromBlock: Int, toBlock: Int, completion: @escaping Completion<[ContractLogEnum]>) {
         
         // Validate historyId
         do {
@@ -73,7 +73,7 @@ final public class ContractsFacadeImp: ContractsFacade, ECHOQueueble {
             return
         }
         
-        services.databaseService.getContractLogs(contractId: contractId, fromBlock: fromBlock, limit: limit, completion: completion)
+        services.databaseService.getContractLogs(contractId: contractId, fromBlock: fromBlock, toBlock: toBlock, completion: completion)
     }
     
     public func getContractResult(contractResultId: String, completion: @escaping Completion<ContractResultEnum>) {
@@ -438,6 +438,7 @@ final public class ContractsFacadeImp: ContractsFacade, ECHOQueueble {
     }
     
     public func queryContract(registrarNameOrId: String,
+                              amount: UInt,
                               assetId: String,
                               contratId: String,
                               methodName: String,
@@ -445,6 +446,7 @@ final public class ContractsFacadeImp: ContractsFacade, ECHOQueueble {
                               completion: @escaping Completion<String>) {
         
         queryContract(registrarNameOrId: registrarNameOrId,
+                      amount: amount,
                       assetId: assetId,
                       contratId: contratId,
                       executeType: ContractExecuteType.nameAndParams(methodName, methodParams),
@@ -452,12 +454,14 @@ final public class ContractsFacadeImp: ContractsFacade, ECHOQueueble {
     }
     
     public func queryContract(registrarNameOrId: String,
+                              amount: UInt,
                               assetId: String,
                               contratId: String,
                               byteCode: String,
                               completion: @escaping Completion<String>) {
         
         queryContract(registrarNameOrId: registrarNameOrId,
+                      amount: amount,
                       assetId: assetId,
                       contratId: contratId,
                       executeType: ContractExecuteType.code(byteCode),
@@ -465,6 +469,7 @@ final public class ContractsFacadeImp: ContractsFacade, ECHOQueueble {
     }
     
     fileprivate func queryContract(registrarNameOrId: String,
+                                   amount: UInt,
                                    assetId: String,
                                    contratId: String,
                                    executeType: ContractExecuteType,
@@ -504,7 +509,7 @@ final public class ContractsFacadeImp: ContractsFacade, ECHOQueueble {
         
         // Operation
         queryQueue.saveValue(Contract(id: contratId), forKey: ContractKeys.receiverContract.rawValue)
-        let callContractNoChangigState = createBildCallContractNoChangingState(queryQueue, assetId, completion)
+        let callContractNoChangigState = createBildCallContractNoChangingState(queryQueue, amount, assetId, completion)
         
         // Completion
         let completionOperation = createCompletionOperation(queue: queryQueue)
@@ -623,6 +628,7 @@ final public class ContractsFacadeImp: ContractsFacade, ECHOQueueble {
     }
     
     fileprivate func createBildCallContractNoChangingState(_ queue: ECHOQueue,
+                                                           _ amount: UInt,
                                                            _ assetId: String,
                                                            _ completion: @escaping Completion<String>) -> Operation {
         let contractOperation = BlockOperation()
@@ -636,6 +642,7 @@ final public class ContractsFacadeImp: ContractsFacade, ECHOQueueble {
             guard let receive: Contract = queue?.getValue(ContractKeys.receiverContract.rawValue) else { return }
             
             self?.services.databaseService.callContractNoChangingState(contract: receive,
+                                                                       amount: amount,
                                                                        asset: Asset(assetId),
                                                                        account: account,
                                                                        contractCode: byteCode,