+++ ECHO/Core/Facades/Implementations/InformationFacadeImp.swift
@@ -583,7 +583,7 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
             
             let depositsIdsArray = depositsIds.map { $0 }
             
-            self?.services.databaseService.getObjects(type: DepositEth.self,
+            self?.services.databaseService.getObjects(type: EthDeposit.self,
                                                       objectsIds: depositsIdsArray,
                                                       completion: { (result) in
                 
@@ -616,7 +616,7 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
             
             let withdrawsIdsArray = withdrawsIds.map { $0 }
             
-            self?.services.databaseService.getObjects(type: WithdrawalEth.self,
+            self?.services.databaseService.getObjects(type: EthWithdrawal.self,
                                                       objectsIds: withdrawsIdsArray,
                                                       completion: { (result) in
                                                         
@@ -744,13 +744,13 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
                     historyItem.operation = operation
                 }
                 
-                if var operation = operation as? SidechainETHIssueOperation {
+                if var operation = operation as? SidechainIssueOperation {
                     let account = self?.findAccountIn(accounts, accountId: operation.account.id)
                     operation.changeAccount(account: account)
                     historyItem.operation = operation
                 }
                 
-                if var operation = operation as? SidechainETHBurnOperation {
+                if var operation = operation as? SidechainBurnOperation {
                     let account = self?.findAccountIn(accounts, accountId: operation.account.id)
                     operation.changeAccount(account: account)
                     historyItem.operation = operation
@@ -773,7 +773,7 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
             
             guard mergeDepositsEthInHistoryOperation?.isCancelled == false else { return }
             guard var history: [HistoryItem] = queue?.getValue(AccountHistoryResultsKeys.historyItems.rawValue) else { return }
-            guard let deposits: [DepositEth] = queue?.getValue(AccountHistoryResultsKeys.loadedDepositsIds.rawValue) else { return }
+            guard let deposits: [EthDeposit] = queue?.getValue(AccountHistoryResultsKeys.loadedDepositsIds.rawValue) else { return }
             
             for index in 0..<history.count {
                 
@@ -781,7 +781,7 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
                 
                 guard let operation = historyItem.operation else { continue }
                 
-                if var operation = operation as? SidechainETHIssueOperation {
+                if var operation = operation as? SidechainIssueOperation {
                     let deposit = self?.findDepositEthIn(deposits, depositId: operation.depositId)
                     operation.deposit = deposit
                     historyItem.operation = operation
@@ -804,7 +804,7 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
             
             guard mergeDepositsEthInHistoryOperation?.isCancelled == false else { return }
             guard var history: [HistoryItem] = queue?.getValue(AccountHistoryResultsKeys.historyItems.rawValue) else { return }
-            guard let withdraws: [WithdrawalEth] = queue?.getValue(AccountHistoryResultsKeys.loadedWithdrawalsIds.rawValue) else { return }
+            guard let withdraws: [EthWithdrawal] = queue?.getValue(AccountHistoryResultsKeys.loadedWithdrawalsIds.rawValue) else { return }
             
             for index in 0..<history.count {
                 
@@ -812,7 +812,7 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
                 
                 guard let operation = historyItem.operation else { continue }
                 
-                if var operation = operation as? SidechainETHBurnOperation {
+                if var operation = operation as? SidechainBurnOperation {
                     let withdraw = self?.findWithdrawsEthIn(withdraws, withdrawId: operation.withdrawId)
                     operation.withdraw = withdraw
                     historyItem.operation = operation
@@ -925,14 +925,14 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
                     historyItem.operation = operation
                 }
                 
-                if var operation = operation as? SidechainETHIssueOperation {
+                if var operation = operation as? SidechainIssueOperation {
                     let feeAsset = self?.findAssetsIn(assets, assetId: operation.fee.asset.id)
                     let valueAsset = self?.findAssetsIn(assets, assetId: operation.value.asset.id)
                     operation.changeAssets(valueAsset: valueAsset, feeAsset: feeAsset)
                     historyItem.operation = operation
                 }
                 
-                if var operation = operation as? SidechainETHBurnOperation {
+                if var operation = operation as? SidechainBurnOperation {
                     let feeAsset = self?.findAssetsIn(assets, assetId: operation.fee.asset.id)
                     let valueAsset = self?.findAssetsIn(assets, assetId: operation.value.asset.id)
                     operation.changeAssets(valueAsset: valueAsset, feeAsset: feeAsset)
@@ -975,12 +975,12 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
         return array.first(where: {$0.id == assetId})
     }
     
-    fileprivate func findDepositEthIn(_ array: [DepositEth], depositId: String) -> DepositEth? {
+    fileprivate func findDepositEthIn(_ array: [EthDeposit], depositId: String) -> EthDeposit? {
         
         return array.first(where: {$0.id == depositId})
     }
     
-    fileprivate func findWithdrawsEthIn(_ array: [WithdrawalEth], withdrawId: String) -> WithdrawalEth? {
+    fileprivate func findWithdrawsEthIn(_ array: [EthWithdrawal], withdrawId: String) -> EthWithdrawal? {
         
         return array.first(where: {$0.id == withdrawId})
     }
@@ -1068,12 +1068,12 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
                 return
             }
             
-            if let operation = operation as? SidechainETHIssueOperation {
+            if let operation = operation as? SidechainIssueOperation {
                 accountsIds.insert(operation.account.id)
                 return
             }
             
-            if let operation = operation as? SidechainETHBurnOperation {
+            if let operation = operation as? SidechainBurnOperation {
                 accountsIds.insert(operation.account.id)
                 return
             }
@@ -1129,12 +1129,12 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
                 return
             }
             
-            if let operation = operation as? SidechainETHIssueOperation {
+            if let operation = operation as? SidechainIssueOperation {
                 assetsIds.insert(operation.value.asset.id)
                 return
             }
             
-            if let operation = operation as? SidechainETHBurnOperation {
+            if let operation = operation as? SidechainBurnOperation {
                 assetsIds.insert(operation.value.asset.id)
                 return
             }
@@ -1153,7 +1153,7 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
                 return
             }
             
-            if let operation = operation as? SidechainETHIssueOperation {
+            if let operation = operation as? SidechainIssueOperation {
                 depositsId.insert(operation.depositId)
                 return
             }
@@ -1172,7 +1172,7 @@ final public class InformationFacadeImp: InformationFacade, ECHOQueueble {
                 return
             }
             
-            if let operation = operation as? SidechainETHBurnOperation {
+            if let operation = operation as? SidechainBurnOperation {
                 withdrawalsId.insert(operation.withdrawId)
                 return
             }