+++ ECHONetworkTests/Subscribe/BlocksSubscribeTests.swift
@@ -40,7 +40,17 @@ class BlocksSubscribeTests: XCTestCase, SubscribeBlockDelegate {
                 
                 self.echo.subscribeToBlock(delegate: self)
                 
-                DispatchQueue.main.asyncAfter(deadline: .now() + 5, execute: {
+                DispatchQueue.main.asyncAfter(deadline: .now() + 2, execute: {
+                    self.echo.sendTransferOperation(fromNameOrId: Constants.defaultName,
+                                                    wif: Constants.defaultWIF,
+                                                    toNameOrId: Constants.defaultToName,
+                                                    amount: 1, asset: Constants.defaultAsset,
+                                                    assetForFee: nil,
+                                                    completion: { _ in},
+                                                    noticeHandler: nil)
+                })
+                
+                DispatchQueue.main.asyncAfter(deadline: .now() + 8, execute: {
                     exp.fulfill()
                 })
             case .failure(let error):